backend/requirements.txt
fastapi == 0.115.8
uvicorn == 0.34.0
SQLAlchemy == 2.0.38
psycopg2 == 2.9.10
requests2 == 0.2